Se realiza la creación de los graficos mediante la consulta de los diferentes estadísticos que se tomaron. Se realizan los respectivos controladores para que hagan la funcion de delegación. Se realizan pruebas unitarias de las clases

+++ src/co/udea/edu/proyectointegrador/gr11/parqueaderoapp/domain/stadistics/entities/TipoUsuarioEstadistica.java
@@ -15,18 +15,18 @@ public class TipoUsuarioEstadistica {
     
     private TipoUsuario tipoUsuario;
      
-    private int numeroDeIngresos;
+    private long numeroDeIngresos;
 
-    public TipoUsuarioEstadistica(TipoUsuario tipoUsuario, int numeroDeIngresos) {
+    public TipoUsuarioEstadistica(TipoUsuario tipoUsuario, long numeroDeIngresos) {
         this.tipoUsuario = tipoUsuario;
         this.numeroDeIngresos = numeroDeIngresos;
     }
     
-    public int getNumeroDeIngresos() {
+    public long getNumeroDeIngresos() {
         return numeroDeIngresos;
     }
 
-    public void setNumeroDeIngresos(int numeroDeIngresos) {
+    public void setNumeroDeIngresos(long numeroDeIngresos) {
         this.numeroDeIngresos = numeroDeIngresos;
     }
